adding in an approach for optimizing a divergent two cluster set of parameter model points. At each gene optimizes relative to a set of models, rather than against a single model

+++ neuronunit/optimization/optimization_management.py
@@ -816,7 +816,7 @@ def filtered(pop,dtcpop):
 
 
 def split_list(a_list):
-    half = len(a_list)/2
+    half = int(len(a_list)/2)
     return a_list[:half], a_list[half:]
 
 def clusty(dtcbag,dtcpop):
